Removes exception handlers from StudentRestController
Eliminates custom exception handler methods for handling
specific and generic exceptions. Likely simplifies the
controller by delegating exception handling to a global
or centralized mechanism, improving maintainability.
